test: update cli e2e test to work with refactored worker
