+++ apps/cli/tests/fixtures/branching-flow-failure.json
@@ -17,15 +17,15 @@
     "second": {
       "type": "action",
       "tool": "stt-client",
-      "op": "transcribe",
+      "op": "echo",
       "args": {
-        "text": "${steps.first.text}"
+        "text": "${steps.transcribe.result.structuredContent.text}"
       }
     },
     "third": {
       "type": "action",
       "tool": "stt-client",
-      "op": "transcribe",
+      "op": "echo",
       "args": {
         "text": "it failed"
       }
